No crashes. Mostly working.

+++ sdsai-common/src/main/java/com/github/basking2/sdsai/marchinesquares/Feature.java
@@ -29,7 +29,7 @@ public class Feature {
         this.properties = new HashMap<>();
     }
 
-    public void translate(final int xOffset, final int yOffset) {
+    public void translate(final double xOffset, final double yOffset) {
         for (final Point point: points) {
             point.x += xOffset;
             point.y += yOffset;
